overlay/081: fix test when running with index enabled
Test overlay/081 fails with:
 CONFIG_OVERLAY_FS_INDEX=y
or
 echo Y > /sys/modules/overlay/params/index

The reason is that mount option uuid=off has the undesired side effect
of disabling index feature.

uuid=null is exactly the same as uuid=off for the purpose of this test
but without the undesired side effect.

The test was created to test the new modes uuid=null/auto/on, so the
fact that is is testing the mode uuid=off is just an oversight.

Covert the use of uuid=off to uuid=null to fix this problem.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>
Reviewed-by: Zorro Lang <zlang@redhat.com>
Signed-off-by: Zorro Lang <zlang@kernel.org>